backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
SQLAlchemy==2.0.23